replace fpSetP* with mxSetP*
The new versions don't seem to work for me (2011b) when returning a
fpAllocate'd array back to matlab - probably the mxArray struct has
changed in this version (and is now hidden). The matlab mxSetPr
functions work fine.
On non-matlab managed memory the mxSetPr gives a runtime exception, but
for me being able to return matlab allocated array from fortran is more
important.